test(compile): Pull out some redundant freshness checks
These check for context independent of mtime vs checksum
